Debug: Extract classes that are not Kore-specific
